Version 1 Almost Done. So, this is not completed, so normally I wouldn't commit it However, with how broken the previous version was I'm just full sending this. I have set up interfaces for moves and gamestates. If you just implement these for the new games you wish to add then everything should just work.
As far as improvements still to be made... We still need wolves vs dogs implemented, Nim Sum calculations should be done for instances where the tree cannot fully be explored as the depth given doesn't reach far enough, and it might be good to look at some performance optimisations for the minimax algorithm.

Overall the code is still pretty unpolished still, but its so so so much better than the previous version.